pyproject.toml
[project]
name = "sbwatch"
version = "0.5.0"
description = "SB Watchbot v5: AI-based Silver Bullet alert bot"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0"
]
[project.scripts]
sbwatch = "sbwatch.cli.main:app"
[tool.setuptools.packages.find]
where = ["src"]
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
